Split module into separate files
