+++ src/AuthContext.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const { Consumer, Provider } = React.createContext('authorizator');
+
+export {
+  Consumer,
+  Provider,
+};
